fix: Add final HTML tag angle bracket only once
For HTML tag inputs that have attributes and contain the final angle
bracket (e.g. `<div class="test">`), insert the final angle bracket only
once.

Previously, the start tag would be `<div class="test">>`. The final
angle bracket would be inserted twice.

This seems to be due to `(.*)` (in the `(.*)>?` regular expression)
lazily including the final `>` of the input **inside** of the `.*`
matching group. By using `[^>]` instead of `.*`, the final angle bracket
is correctly matched by `>?`.