tests: Use a consistent shiftwidth in tests. (#81)
When running the tests using `<Plug>PlenaryTestFile` the `shiftwidth`
option was inherited from the user's config. For users that modified
`shiftwidth` in their configuration the tests were failing. One of the
tests ("nvim-surround can visual-line surround") indented a line while
doing a surround but due to a different `shiftwidth`, the size of the
indent did not match the expected 4 spaces.

It is possible that other indent-related options should also be set so
the tests behave consistently. `expandtab` comes to mind.