fix: Correctly apply indent in insert_surround()
Fixes indent when 'shiftwidth' is 0, 'vartabstop' is set, or
'shiftwidth' is not a multiple of 'tabstop'.

shiftwidth() (as opposed to the 'shiftwidth' option directly) gives the
effective shiftwidth, returning the first value of 'vartabstop' if that
option is set, else 'shiftwidth' if it is non-zero, else 'tabstop'. This
matches what the = operator does.