feat: Implement `change_line` mapping.
Resolves #244.

* Implement preliminary change_line mapping

As of this commit, given the input, where `|` denotes the cursor
position:
```
let x = foo(args|);
```

`cS)(` yields:
```
let x = foo(|
    args
);
```

The indentation of `args` depends on the indentation rules for the filetype.

We still need to iron out behavior for complex surrounds like HTML tags
and functions. Currently, given the input, where `|` denotes the cursor
position:
```
let x = foo(args|);
```

Both `csf` AND `cSf` with `bar` would yield:
```
let x = |bar(args);
```

This follows because the change target selection is just the `foo`.
However, users may instead want the result to be:
```
let x = |bar(
    args
);
```

This is more useful because otherwise, `cs` and `cS` would be exhibiting
the same behavior.

* Write basic tests for change_line

* Add missing `delimiters` type annotation for normal cache

This is unrelated to the PR, but Kyle caught this while reviewing it and
asked me to put it in.