fix: Translate char to termcodes when reading input (#271)
In ‘config.translate_opts()’, the chars specified in the ‘surrounds’
portion of a configuration are translated to Vim internal representation
in order to support special characters.  As a result though, we need to
actually translate the users input.  Otherwise configuration of certain
Unicode characters such as U+2019 RIGHT SINGLE QUOTATION MARK may not
work.