fix(dot-repeat): Clear dot-repeat after indentation (#306)
As it stands, `indent_lines` clears the dot-repeat behavior without 
clearing the plugin cache, and so successive `.` after a command that adds new
lines (e.g. `ySS`) will merely re-indent the lines without doing anything else.
This change forces the error `'operatorfunc' is empty`, which doesn't pollute
the undo tree, allowing the user to simply hit `u` once to revert the surround.

Ideally, we would repeat the surround, but the current implementation of
`indent_lines` wipes the current dot-repeat action, which makes this
infeasible to implement.